dataset to long

+++ llm_non_identifiability/dataset.py
@@ -9,7 +9,7 @@ from llm_non_identifiability.data import pad
 class GrammarDataset(Dataset):
     def __init__(self, data):
         # pad the data
-        self.data = torch.from_numpy(pad(data))
+        self.data = torch.from_numpy(pad(data)).long()
         self.labels = self.data.clone()
 
     def __len__(self):
